Split multitenant-specific startup into separate file
Get rid of explicit type-checking in common.go so that we can use
this code for other plugins too, later.
